% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dpotential}
\alias{dpotential}
\title{Compute derivative of potential function}
\usage{
dpotential(x = seq(0, 2, l = 100), a, r, H, Q, K)
}
\arguments{
\item{x}{}

\item{a}{}

\item{r}{}

\item{H}{}

\item{Q}{}

\item{K}{}
}
\value{

}
\description{
Compute derivative of potential function
}
