% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_mcmc}
\alias{fit_mcmc}
\title{Fit parametric model using nimble}
\usage{
fit_mcmc(
  data,
  constants,
  inits = NULL,
  seed,
  n_iterations = 10,
  MVsampler_type = "AF_slice"
)
}
\arguments{
\item{data}{}

\item{constants}{}

\item{inits}{}

\item{seed}{}

\item{n_iterations}{}

\item{MVsampler_type}{}
}
\value{

}
\description{
Fit parametric model using nimble
}
