% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_mcmc_functional}
\alias{fit_mcmc_functional}
\title{Fit functional model using nimble}
\usage{
fit_mcmc_functional(
  data,
  constants,
  inits = NULL,
  seed,
  n_iterations = 10,
  MVsampler_type = "AF_slice"
)
}
\arguments{
\item{data}{}

\item{constants}{}

\item{inits}{}

\item{seed}{}

\item{n_iterations}{in MCMC algorithm}

\item{MVsampler_type}{either "AF_slice" or "RW_block"}
}
\value{
list including samples from MCMC, simulated data, and timing information
}
\description{
Fit functional model using nimble
}
