% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_ISE}
\alias{get_ISE}
\title{Extract ISE for parametric model}
\usage{
get_ISE(samples, true, x = seq(0, 2, length.out = 100))
}
\arguments{
\item{samples}{}

\item{true}{}

\item{x}{}
}
\value{
scalar; integrated squared error between true derivative of potential and posterior distribution
}
\description{
Extract ISE for parametric model
}
