% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_ISE_functional}
\alias{get_ISE_functional}
\title{Extract ISE for functional model}
\usage{
get_ISE_functional(samples, true, x = seq(0, 2, length.out = 100))
}
\arguments{
\item{samples}{}

\item{true}{}

\item{x}{}
}
\value{

}
\description{
Extract ISE for functional model
}
