\name{longTransients-package}
\alias{longTransients-package}
\alias{longTransients}
\docType{package}
\title{
\packageTitle{longTransients}
}
\description{
\packageDescription{longTransients}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{longTransients}
\packageIndices{longTransients}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{longTransients}

Maintainer: \packageMaintainer{longTransients}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
