% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{make_model_code_functional_nC}
\alias{make_model_code_functional_nC}
\title{Create nimbleCode object with functional model}
\usage{
make_model_code_functional_nC()
}
\value{
nimbleCode object
}
\description{
Create nimbleCode object with functional model
}
