% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{Traceplots for MCMC output from parameteric model}
\usage{
plot_trace(samples, true = NULL)
}
\arguments{
\item{samples}{}

\item{true}{values used to generate data for comparison}
}
\description{
Traceplots for MCMC output from parameteric model
}
