% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace_functional}
\alias{plot_trace_functional}
\title{Traceplots for MCMC output from functional model}
\usage{
plot_trace_functional(samples)
}
\arguments{
\item{samples}{}
}
\description{
Traceplots for MCMC output from functional model
}
