% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot population trajectories}
\usage{
plot_traj(obs_y = NULL, true_x = NULL, t.step = 1)
}
\arguments{
\item{obs_y}{observed population with error}

\item{true_x}{true population}

\item{t.step}{time step}
}
\description{
Plot population trajectories
}
