% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_potential_functional}
\alias{plot_potential_functional}
\title{Plot realizations of the potential function and its derivative for the parametric model}
\usage{
plot_potential_functional(
  samples,
  n_subset = 400,
  true = NULL,
  obs_y = NULL,
  x = seq(0, 2, length.out = 100),
  probs = c(0.125, 0.5, 0.875),
  alpha = 0.02,
  ylim_potential = NULL,
  ylim_dpotential = NULL
)
}
\arguments{
\item{samples}{}

\item{n_subset}{}

\item{true}{}

\item{obs_y}{observations}

\item{x}{}

\item{probs}{}

\item{alpha}{}

\item{ylim_potential}{}

\item{ylim_dpotential}{}
}
\value{

}
\description{
Plot realizations of the potential function and its derivative for the parametric model
}
